Cleanup entry::mode API
