+++ gix-index/src/entry/mode.rs
@@ -1,36 +1,15 @@
-use bitflags::bitflags;
-
-// TODO: we essentially treat this as an enum withj the only exception being
-// that `FILE_EXECUTABLE.contains(FILE)` works might want to turn this into an
-// enum proper
-bitflags! {
-    /// The kind of file of an entry.
-    #[derive(Copy, Clone, Debug, PartialEq, Eq)]
-    pub struct Mode: u32 {
-        /// directory (only used for sparse checkouts), equivalent to a tree, which is _excluded_ from the index via
-        /// cone-mode.
-        const DIR = 0o040000;
-        /// regular file
-        const FILE = 0o100644;
-        /// regular file, executable
-        const FILE_EXECUTABLE = 0o100755;
-        /// Symbolic link
-        const SYMLINK = 0o120000;
-        /// A git commit for submodules
-        const COMMIT = 0o160000;
-    }
-}
+use crate::entry::Mode;
 
 #[cfg(unix)]
 /// Returns whether a a file has the executable permission set
-pub fn is_executable(metadata: &std::fs::Metadata) -> bool {
+fn is_executable(metadata: &std::fs::Metadata) -> bool {
     use std::os::unix::fs::MetadataExt;
     (metadata.mode() & 0o100) != 0
 }
 
 #[cfg(not(unix))]
 /// Returns whether a a file has the executable permission set
-pub fn is_executable(_metadata: &std::fs::Metadata) -> bool {
+fn is_executable(_metadata: &std::fs::Metadata) -> bool {
     false
 }
 
@@ -80,11 +59,16 @@ pub enum Change {
 }
 
 impl Change {
-    /// Applies this change to a `Mode`
-    pub fn apply(self, mode: &mut Mode) {
-        *mode = match self {
+    /// Applies this change to a `Mode` by updating it in place
+    pub fn update(self, mode: &mut Mode) {
+        *mode = self.apply(*mode);
+    }
+
+    /// Applies this change to a `Mode` by updating it in place
+    pub fn apply(self, mode: Mode) -> Mode {
+        match self {
             Change::Type { new_mode } => new_mode,
-            Change::ExecutableBit => match *mode {
+            Change::ExecutableBit => match mode {
                 Mode::FILE => Mode::FILE_EXECUTABLE,
                 Mode::FILE_EXECUTABLE => Mode::FILE,
                 _ => unreachable!("invalid mode change: can't flip executable bit of {mode:?}"),
