+++ gix-index/tests/index/entry.rs
@@ -319,3 +319,21 @@ mod stat {
         );
     }
 }
+
+mod mode {
+    use gix_index::entry::mode::Change;
+    use gix_index::entry::Mode;
+
+    #[test]
+    fn apply_change() {
+        assert_eq!(Change::ExecutableBit.apply(Mode::FILE), Mode::FILE_EXECUTABLE);
+        assert_eq!(Change::ExecutableBit.apply(Mode::FILE_EXECUTABLE), Mode::FILE);
+        assert_eq!(
+            Change::Type {
+                new_mode: Mode::SYMLINK
+            }
+            .apply(Mode::FILE),
+            Mode::SYMLINK
+        );
+    }
+}
